requirements.txt
pillow
pymupdf
pypdf
